COMMUNICATION-3413 EI7 LoadProfileXBlockingIssue errors seen in several conditions - correction and imports refactor

+++ upl/com.energyict.mdc.upl.impl/protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/dlms/a2/events/A2MetrologicalEventLog.java
@@ -21,7 +21,7 @@ public class A2MetrologicalEventLog extends AbstractEvent {
         Date eventTimeStamp;
         for (int i = 0; i <= (size - 1); i++) {
             int eventId = (int) this.dcEvents.getRoot().getStructure(i).getValue(1) & 0xFF; // To prevent negative values
-            long unixTime = (long) dcEvents.getRoot().getStructure(i).getInteger(0);
+            long unixTime = (long) dcEvents.getRoot().getStructure(i).getLong(0);
             eventTimeStamp = new Date(unixTime * 1000);
             buildMeterEvent(meterEvents, eventTimeStamp, eventId);
         }
